Move general helpers to specific helpers
